+++ src/grepros/search.py
@@ -69,13 +69,13 @@ class Searcher(object):
             d.clear()
         source.bind(sink), sink.bind(source)
 
-        counter, total, batch = 0, 0, None
+        counter, total, batch, any_matched = 0, 0, None, False
         for topic, msg, stamp in source.read():
             if batch != source.get_batch():
                 total += sum(x[True] for x in self._counts.values())
                 for d in (self._counts, self._messages, self._stamps, self._statuses, self._hashes):
                     d.clear()
-                counter, batch = 0, source.get_batch()
+                counter, batch, any_matched = 0, source.get_batch(), False
 
             msgid = counter = counter + 1
             topickey = (topic, rosapi.get_message_type(msg))
@@ -100,10 +100,12 @@ class Searcher(object):
                     sink.emit(topic, i, s, m, None)
                     self._statuses[topickey][msgid2] = False
             source.notify(matched)
+            any_matched = any_matched or bool(matched)
 
             self._prune_data(topickey)
-            if self._is_max_done(source):
-                break  # for topic, msg, stamp
+            if any_matched and self._is_max_done(source):
+                sink.flush()
+                source.close_batch()
 
         source.close(), sink.close()
         return total + sum(x[True] for x in self._counts.values())
@@ -186,20 +188,17 @@ class Searcher(object):
 
     def _is_max_done(self, source):
         """Returns whether max match count has been reached (and message after-context emitted)."""
-        result, is_maxed, is_maxable = False, False, False
+        result, is_maxed = False, False
         if self._args.MAX_MATCHES:
             is_maxed = sum(vv[True] for vv in self._counts.values()) >= self._args.MAX_MATCHES
         if not is_maxed and self._args.MAX_TOPIC_MATCHES:
-            if self._args.MAX_TOPICS:
-                if len([k for k, vv in self._counts.items() if vv[True]]) >= self._args.MAX_TOPICS:
-                    is_maxable = True
-            else:
-                is_maxable = source.is_static
-            is_maxed = is_maxable and any(vv[True] for vv in self._counts.values()) and \
-                       all(vv[True] >= self._args.MAX_TOPIC_MATCHES
-                           for vv in self._counts.values() if vv[True])
+            count_required = self._args.MAX_TOPICS or len(source.topics)
+            count_maxed = sum(vv[True] >= self._args.MAX_TOPIC_MATCHES
+                              for vv in self._counts.values())
+            is_maxed = (count_maxed >= count_required)
         if is_maxed:
-            result = not any(self._has_in_window(k, self._args.AFTER, status=True, full=True)
+            result = not self._args.AFTER or \
+                     not any(self._has_in_window(k, self._args.AFTER, status=True, full=True)
                              for k in self._counts)
         return result
 
@@ -239,7 +238,7 @@ class Searcher(object):
         def decorate_message(obj, top=()):
             """Recursively converts field values to pattern-matched strings."""
             selects, noselects = self._patterns["select"], self._patterns["noselect"]
-            fieldmap = rosapi.get_message_fields(obj)  # Returns obj if not a ROS message
+            fieldmap = rosapi.get_message_fields(obj)  # Returns obj if not ROS message
             if fieldmap != obj:
                 fieldmap = filter_fields(fieldmap, top, include=selects, exclude=noselects)
             for k, t in fieldmap.items() if fieldmap != obj else ():